// src/types/onyx/ReportAction.ts
type Message = {
    /** The type of the action item fragment. Used to render a corresponding component */
    type: string;

    /** The text content of the fragment. */
    text: string;

    /** The html content of the fragment. */
    html?: string;

    /** Used to apply additional styling. Style refers to a predetermined constant and not a class name. e.g. 'normal'
     * or 'strong'
     */
    style?: string;

    /** ID of a report */
    reportID?: string;

    /** ID of a policy */
    policyID?: string;

    /** The target of a link fragment e.g. '_blank' */
    target?: string;

    /** The destination of a link fragment e.g. 'https://www.ieatta.com' */
    href?: string;

    /** An additional avatar url - not the main avatar url but used within a message. */
    iconUrl?: string;

    /** Fragment edited flag */
    isEdited?: boolean;

    /** Whether thread's parent message is deleted or not */
    isDeletedParentAction?: boolean;

    /** Whether the pending transaction was reversed and didn't post to the card */
    isReversedTransaction?: boolean;
    whisperedTo?: number[];
    reactions?: Reaction[];

    moderationDecision?: Decision;
    translationKey?: string;

    /** ID of a task report */
    taskReportID?: string;

    /** Reason of payment cancellation */
    cancellationReason?: string;

    /** ID of an expense report */
    expenseReportID?: string;

    /** Amount of an expense */
    amount?: number;

    /** Currency of an expense */
    currency?: string;

    /** resolution for actionable mention whisper */
    resolution?: ValueOf<typeof CONST.REPORT.ACTIONABLE_MENTION_WHISPER_RESOLUTION> | null;

    /** The time this report action was deleted */
    deleted?: string;
};